+++ ezdxf/math/ucs.py
@@ -1,7 +1,8 @@
-# Copyright (c) 2018 Manfred Moitzi
+# Copyright (c) 2018-2019 Manfred Moitzi
 # License: MIT License
 from typing import TYPE_CHECKING, Tuple, Sequence, Iterable
 from .vector import Vector, X_AXIS, Y_AXIS, Z_AXIS
+from .matrix44 import Matrix44
 
 if TYPE_CHECKING:
     from ezdxf.eztypes import GenericLayoutType, Vertex, BaseLayout
@@ -20,6 +21,7 @@ class Matrix33:
     Simple 3x3 Matrix for coordinate transformation.
 
     """
+    # faster transformation than Matrix44
     __slots__ = ('ux', 'uy', 'uz')
 
     def __init__(self, ux: 'Vertex' = (1, 0, 0), uy: 'Vertex' = (0, 1, 0), uz: 'Vertex' = (0, 0, 1)):
@@ -242,6 +244,22 @@ class UCS:
         """ Returns iterable of UCS vectors from WCS `points`. """
         for point in points:
             yield self.from_wcs(point)
+
+    def rotate(self, axis: 'Vertex', angle: float) -> 'UCS':
+        """
+        Returns a new rotated UCS, with the same origin as the source UCS.
+        The rotation vector is located in the origin and has :ref:`WCS` coordinates.
+
+        .. versionadded:: 0.11
+
+        Args:
+            axis: arbitrary rotation axis as vector in :ref:`WCS`
+            angle: rotation angle in radians
+
+        """
+        t = Matrix44.axis_rotate(Vector(axis), angle)
+        ux, uy, uz = t.transform_vectors([self.ux, self.uy, self.uz])
+        return UCS(origin=self.origin, ux=ux, uy=uy, uz=uz)
 
     @property
     def is_cartesian(self) -> bool: