+++ NEWS.md
@@ -16,6 +16,7 @@ Version 0.11a0 - dev
 - NEW: `Drawing.set_modelspace_vport(height, center)` set initial view/zoom location for the modelspace
 - NEW: 2d construction function `ezdxf.math.offset_vertices_2d()`
 - NEW: `Drawing.output_encoding`  returns required output encoding
+- NEW: `UCS.rotate(axis, angle)` returns a new rotated UCS
 - BUGFIX: fixed base point calculation of aligned dimensions
 - BUGFIX: fixed length extension line support